refactor(perps): clean up `/markets/` files as they're no longer used
